+++ tests/models/multimodal/encoders/test_encoders2d.py
@@ -1,9 +1,22 @@
-from pvnet.models.multimodal.encoders.encoders3d import (
-    DefaultPVNet,
-    DefaultPVNet2,
-    EncoderUNET,
+from pvnet.models.multimodal.encoders.encoders2d import (
+    NaiveEfficientNet,
+    NaiveResNet,
+    ConvNeXt,
+    CNBlockConfig,
 )
 import pytest
+
+
+@pytest.fixture()
+def convnext_model_kwargs(encoder_model_kwargs):
+    model_kwargs = {k: v for k, v in encoder_model_kwargs.items()}
+    model_kwargs["block_setting"] = [
+        CNBlockConfig(96, 192, 3),
+        CNBlockConfig(192, 384, 3),
+        CNBlockConfig(384, 768, 9),
+        CNBlockConfig(768, None, 3),
+    ]
+    return model_kwargs
 
 
 def _test_model_forward(batch, model_class, model_kwargs):
@@ -20,26 +33,30 @@ def _test_model_backward(batch, model_class, model_kwargs):
 
 
 # Test model forward on all models
-def test_defaultpvnet_forward(sample_satellite_batch, encoder_model_kwargs):
-    _test_model_forward(sample_satellite_batch, DefaultPVNet, encoder_model_kwargs)
+def test_naiveefficientnet_forward(sample_satellite_batch, encoder_model_kwargs):
+    # Skip if optional dependency not installed
+    pytest.importorskip("efficientnet_pytorch")
+    _test_model_forward(sample_satellite_batch, NaiveEfficientNet, encoder_model_kwargs)
 
 
-def test_defaultpvnet2_forward(sample_satellite_batch, encoder_model_kwargs):
-    _test_model_forward(sample_satellite_batch, DefaultPVNet2, encoder_model_kwargs)
+def test_naiveresnet_forward(sample_satellite_batch, encoder_model_kwargs):
+    _test_model_forward(sample_satellite_batch, NaiveResNet, encoder_model_kwargs)
 
 
-def test_encoderunet_forward(sample_satellite_batch, encoder_model_kwargs):
-    _test_model_forward(sample_satellite_batch, EncoderUNET, encoder_model_kwargs)
+def test_convnext_forward(sample_satellite_batch, convnext_model_kwargs):
+    _test_model_forward(sample_satellite_batch, ConvNeXt, convnext_model_kwargs)
 
 
 # Test model backward on all models
-def test_defaultpvnet_backward(sample_satellite_batch, encoder_model_kwargs):
-    _test_model_backward(sample_satellite_batch, DefaultPVNet, encoder_model_kwargs)
+def test_naiveefficientnet_backward(sample_satellite_batch, encoder_model_kwargs):
+    # Skip if optional dependency not installed
+    pytest.importorskip("efficientnet_pytorch")
+    _test_model_backward(sample_satellite_batch, NaiveEfficientNet, encoder_model_kwargs)
 
 
-def test_defaultpvnet2_backward(sample_satellite_batch, encoder_model_kwargs):
-    _test_model_backward(sample_satellite_batch, DefaultPVNet2, encoder_model_kwargs)
+def test_naiveresnet_backward(sample_satellite_batch, encoder_model_kwargs):
+    _test_model_backward(sample_satellite_batch, NaiveResNet, encoder_model_kwargs)
 
 
-def test_encoderunet_backward(sample_satellite_batch, encoder_model_kwargs):
-    _test_model_backward(sample_satellite_batch, EncoderUNET, encoder_model_kwargs)
+def test_convnext_backward(sample_satellite_batch, convnext_model_kwargs):
+    _test_model_backward(sample_satellite_batch, ConvNeXt, convnext_model_kwargs)
